requirements.txt
langchain==0.3.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-groq==0.2.0
langchain-text-splitters==0.3.5
streamlit==1.38.0
python-dotenv==1.0.1
google-cloud-texttospeech==2.11.0


